解决cpu加载量化模型运行报精度不匹配的问题
使用量化后的模型chatglm2-6b-int4启动项目后，报错：expected scalar type Double but found Float。这是因为models\loader\loader.py加载模型时给的参数精度不对，默认用torch.float在GPU上为32位的精度，但是在CPU上为64位的精度，改成torch.float32即可。

+++ models/loader/loader.py
@@ -256,7 +256,7 @@ class LoaderCheckPoint:
             print(
                 "Warning: self.llm_device is False.\nThis means that no use GPU  bring to be load CPU mode\n")
             params = {"low_cpu_mem_usage": True, "torch_dtype": torch.float32, "trust_remote_code": True}
-            model = LoaderClass.from_pretrained(checkpoint, **params).to(self.llm_device, dtype=float)
+            model = LoaderClass.from_pretrained(checkpoint, **params).to(self.llm_device, dtype=torch.float32)
 
         # Loading the tokenizer
         if type(model) is transformers.LlamaForCausalLM:
